+++ client/src/components/DashboardTop/FriendsListItem.js
