+++ client/src/components/DashboardTop/index.js
@@ -1,0 +1,1 @@
+export { default } from "./DashboardTop";
