+++ client/src/pages/Profile.js
@@ -1,17 +1,17 @@
 import React from "react";
 
 import DashboardLayout from "../components/DashboardLayout";
-import Profile from "../components/PublicProfile/ProfilePage";
+import PublicProfile from "../components/PublicProfile";
 import { NAVBAR_HEIGHT } from "../utils/constants";
 
-export default function Friends() {
+export default function Profile() {
   return (
     <DashboardLayout
       style={{
         height: `calc(100vh - ${NAVBAR_HEIGHT})`,
       }}
     >
-      <Profile />
+      <PublicProfile />
     </DashboardLayout>
   );
 }
